+++ weblate/addons/cleanup.py
@@ -20,17 +20,10 @@
 from django.utils.translation import gettext_lazy as _
 
 from weblate.addons.base import UpdateBaseAddon
+from weblate.addons.events import EVENT_POST_COMMIT, EVENT_POST_UPDATE
 
 
 class BaseCleanupAddon(UpdateBaseAddon):
-    @staticmethod
-    def iterate_translations(component):
-        yield from (
-            translation
-            for translation in component.translation_set.iterator()
-            if not translation.is_source or component.intermediate
-        )
-
     @classmethod
     def can_install(cls, component, user):
         if not component.has_template():
@@ -52,3 +45,19 @@ class CleanupAddon(BaseCleanupAddon):
         for translation in self.iterate_translations(component):
             filenames = translation.store.cleanup_unused()
             self.extra_files.extend(filenames)
+
+
+class RemoveBlankAddon(BaseCleanupAddon):
+    name = "weblate.cleanup.blank"
+    verbose = _("Remove blank strings")
+    description = _("Removes strings without a translation from translation files.")
+    events = (EVENT_POST_COMMIT, EVENT_POST_UPDATE)
+    icon = "eraser.svg"
+
+    def update_translations(self, component, previous_head):
+        for translation in self.iterate_translations(component):
+            filenames = translation.store.cleanup_blank()
+            self.extra_files.extend(filenames)
+
+    def post_commit(self, component):
+        self.post_update(component, None, skip_push=True)